Move AccountManagerAsh to chromeos/components/account_manager
Moves AccountManagerAsh to chromeos/components/account_manager to
prepare the upcoming merge of AccountManagerFacade implementations.

This CL moves AccountManagerAsh's ownership from AshChromeServiceImpl to
chromeos::AccountManagerFactory. Subsequent CLs will change
account_manager_facade_factory_ash.cc to instantiate
AccountManagerFacadeImpl instead of AccountManagerFacadeAsh (which will
be removed).

Bug: 1161699
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2616353
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#843723}

+++ chromeos/components/account_manager/account_manager_ash_unittest.cc
@@ -3,7 +3,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "chrome/browser/chromeos/crosapi/account_manager_ash.h"
+#include "chromeos/components/account_manager/account_manager_ash.h"
 
 #include <cstddef>
 #include <memory>
@@ -13,6 +13,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/run_loop.h"
 #include "base/test/bind.h"
 #include "base/test/task_environment.h"
+#include "chromeos/components/account_manager/account_manager.h"
 #include "chromeos/crosapi/mojom/account_manager.mojom-test-utils.h"
 #include "chromeos/crosapi/mojom/account_manager.mojom.h"
 #include "components/account_manager_core/account_manager_util.h"
@@ -117,7 +118,7 @@ class AccountManagerAshTest : public ::testing::Test {
     return account_manager_.IsInitialized();
   }
 
-  size_t GetNumObservers() { return account_manager_ash_->observers_.size(); }
+  int GetNumObservers() { return account_manager_ash_->observers_.size(); }
 
   mojom::AccountManagerAsyncWaiter* account_manager_async_waiter() {
     return account_manager_async_waiter_.get();
